Add files via upload
The Harry Potter Fandom on AO3: A short analyzes of the 500 most popular fanfictions.
Archive of Our Own is, as the name says, an archive that hosts fan works of the most variable fandom on the internet. It is, currently, the most popular site to host fanfictions. In this project, I collect and analyze the data of the most popular fanfics in the Harry Potter fandom. To make the cut for the most popular fanfics I considered only finished works and organized them by Kudos (likes), then taking the first 500 works.
The objective is to use Python language to scrape and analyze the necessary data and create graphics for visualization of the information extracted.